9736: Namespace Sms::MessagesSearcher
